Move asymmetric models to the common syncSchema
